[#80] Refactored insert bookmark dialog to InserBookmarDialog class
